Type:fix
Description:fix scale bug

RootCause:Null

Solution:

+++ application/app/database.cpp
@@ -95,7 +95,7 @@ bool Database::readOperation(DocSheet *sheet)
             sheet->m_operation.mouseShape = (Dr::MouseShape)query.value("mouseShape").toInt();
             sheet->m_operation.scaleMode = (Dr::ScaleMode)query.value("scaleMode").toInt();
             sheet->m_operation.rotation = (Dr::Rotation)query.value("rotation").toInt();
-            sheet->m_operation.scaleFactor = query.value("scaleFactor").toInt();
+            sheet->m_operation.scaleFactor = query.value("scaleFactor").toDouble();
             sheet->m_operation.sidebarVisible = query.value("sidebarVisible").toInt();
             sheet->m_operation.sidebarIndex = query.value("sidebarIndex").toInt();
             sheet->m_operation.currentPage = query.value("currentPage").toInt();
